requirements.txt
Django==1.6.5
MySQL-python==1.2.5
django-admin-tools==0.5.1
#django-autocomplete-light==0.2

flup==1.0.2
pymorphy==0.5.6
simplejson==3.5.3
wsgiref==0.1.2

#django-autocomplete==0.2
https://bitbucket.org/sinyawskiy/django-autocomplete
